pyproject.toml
[project]
name = "adk-bean"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "google-adk>=1.7.0",
    "lxml>=6.0.0",
    "pytrends>=4.9.2",
    "pandas>=2.0.0",
    "requests>=2.25.0",
    "urllib3>=1.26.0",
]
